Add myerror() and use it consistently.
